requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2024.12.13
PyNaCl>=1.5.0 